Removed testing attributes.
